refactor: 불필요한 코드 삭제 및 중복 코드 수정
