+++ BoxOffice/View/BoxOfficeIconCell.swift
@@ -149,37 +149,4 @@ final class BoxOfficeIconCell: UICollectionViewCell {
         configureMovieNameLabel(data: boxOfficeItem)
         configureAudienceCountLabel(data: boxOfficeItem)
     }
-    
-    enum RankInfo {
-        case new
-        case noChange
-        case increase(Int)
-        case decrease(Int)
-        
-        var description: String {
-            switch self {
-            case .new:
-                return "신작"
-            case .noChange:
-                return "-"
-            case .increase(let variance):
-                return "▲\(variance)"
-            case .decrease(let variance):
-                return "▼\(variance * -1)"
-            }
-        }
-        
-        var fontColor: UIColor {
-            switch self {
-            case .new:
-                return .systemRed
-            case .noChange:
-                return .black
-            case .increase:
-                return .systemRed
-            case .decrease:
-                return .systemBlue
-            }
-        }
-    }
 }
